backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
redis==5.0.7
numpy==1.26.4
faiss-cpu==1.8.0.post1
pypdf==4.2.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.2
ruff==0.5.0
